+++ pyshacl/__init__.py
@@ -1,11 +1,13 @@
 # -*- coding: latin-1 -*-
 #
+from .entrypoints import shacl_rules, validate
+from .rule_expand_runner import RuleExpandRunner
 from .shape import Shape
 from .shapes_graph import ShapesGraph
-from .validate import Validator, validate
+from .validator import Validator
 
 # version compliant with https://www.python.org/dev/peps/pep-0440/
 __version__ = '0.26.0'
 # Don't forget to change the version number in pyproject.toml, Dockerfile, and CITATION.cff along with this one
 
-__all__ = ['validate', 'Validator', '__version__', 'Shape', 'ShapesGraph']
+__all__ = ['validate', 'shacl_rules', 'Validator', 'RuleExpandRunner', '__version__', 'Shape', 'ShapesGraph']