+++ pyshacl/run_type.py
@@ -1,0 +1,9 @@
+from abc import ABCMeta, abstractmethod
+
+
+class PySHACLRunType(metaclass=ABCMeta):
+    __slots__ = ()
+
+    @abstractmethod
+    def run(self):
+        raise NotImplementedError()  # pragma: no cover